build: Add optional shell completions feature using clap_complete
This package is very slow, so it's better to make it for now optional feature.

Perhaps I can improve the speed of compilation with a pre-generated completion only for zsh and bash stored in text format on the SCM and then pull it once the user executes the completion they need. But for now I'll just make this as an optional feature to speed things up, because completion is not a critical feature.

After investigating the issue it seems like it takes really long to compile due to intensive macros and codegen in this repo and they do not make it as an optional feature to choose from what completion is needed for what shell, i.e you have to include everything - which is bad.

Signed-off-by: Eden Reich <eden.reich@gmail.com>